Split RequestOptions into RequestOptions/BaseRequestOptions.
